+++ packages/@avatars/jdenticon/src/index.ts
@@ -1,27 +1,29 @@
-import Random from '@avatars/core/lib/random';
+import { IStyle } from '@avatars/core';
 // @ts-ignore
 import jdenticon from 'jdenticon';
 
 type Options = {
   hues?: number[];
-  colorLightness?: [number, number];
-  grayscaleLightness?: [number, number];
-  colorSaturation?: number;
-  grayscaleSaturation?: number;
+  lightnessColor?: [number, number];
+  lightnessGrayscale?: [number, number];
+  saturationColor?: number;
+  saturationGrayscale?: number;
 };
 
-export default function (random: Random, options: Options = {}) {
+const style: IStyle<Options> = function (random, options = {}) {
   jdenticon.config = {
     hues: options.hues,
     lightness: {
-      color: options.colorLightness,
-      grayscale: options.grayscaleLightness,
+      color: options.lightnessColor,
+      grayscale: options.lightnessGrayscale,
     },
     saturation: {
-      color: options.colorSaturation,
-      grayscale: options.grayscaleSaturation,
+      color: options.saturationColor,
+      grayscale: options.saturationGrayscale,
     },
   };
 
   return jdenticon.toSvg(random.seed, 50, 0).replace('width="50"', '').replace('height="50"', '');
-}
+};
+
+export default style;